Fix ParryScore

+++ Assets/Scripts/Player/PlayerParryAction.cs
@@ -92,12 +92,12 @@ public class PlayerParryAction : MonoBehaviour
     private void ParryJudgement()
     {
         if (parryJudgeTime <= goodJudgeTime)
-            GameDataManager.AddDecisionNum(GameDataManager.SCORE_TYPE.GOOD);
+            GameDataManager.AddDecisionNum((int)GameDataManager.SCORE_TYPE.GOOD);
         else
         if (parryJudgeTime <= greatJudgeTime)
-            GameDataManager.AddDecisionNum(GameDataManager.SCORE_TYPE.GREAT);
+            GameDataManager.AddDecisionNum((int)GameDataManager.SCORE_TYPE.GREAT);
         else
         if (parryJudgeTime <= fastJudgeTime)
-            GameDataManager.AddDecisionNum(GameDataManager.SCORE_TYPE.GOOD);
+            GameDataManager.AddDecisionNum((int)GameDataManager.SCORE_TYPE.GOOD);
     }
 }
